pyproject.toml
[tool.poetry]
name = "airquality"
version = "0.1.0"
description = ""
authors = ["Roland Siegbert <roland@siegbert.info>"]

[tool.poetry.dependencies]
python = "3.8"
pyserial = "^3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
